% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catgraph.R
\name{as_ggraph}
\alias{as_ggraph}
\title{Coerce a catgraph to a ggraph-compatible tbl_graph}
\usage{
as_ggraph(x, ...)
}
\arguments{
\item{x}{A \code{catgraph} object.}

\item{...}{Ignored.}
}
\value{
A \code{tbl_graph} object (requires \pkg{tidygraph}).
}
\description{
Coerce a catgraph to a ggraph-compatible tbl_graph
}
