% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catgraph.R
\name{as_igraph}
\alias{as_igraph}
\title{Extract the underlying igraph object from a catgraph}
\usage{
as_igraph(x, ...)
}
\arguments{
\item{x}{A \code{catgraph} object.}

\item{...}{Ignored.}
}
\value{
An \pkg{igraph} undirected graph.
}
\description{
Extract the underlying igraph object from a catgraph
}
